deplace menu test to the end of test suite cause it seems that the deconnexion menu button has some side effects on useRole
